dev-requirements.txt
addonfactory-splunk-conf-parser-lib==0.4.0
antlr4-python3-runtime==4.9.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
dunamai==1.21.2
exceptiongroup==1.2.2
idna==3.7
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
Jinja2==3.1.4
jsonschema==4.17.3
MarkupSafe==2.1.5
openapi3==1.8.2
packaging==24.0
pkgutil_resolve_name==1.3.10
pluggy==1.2.0
pyrsistent==0.19.3
pytest==7.4.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
splunk-add-on-ucc-framework==5.47.0
stix2==3.0.1
stix2-patterns==2.0.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.19
zipp==3.15.0
